Fixes for better error handling.
Changed `bundle()` method to return a promise that resolves when
bundling finishes, and rejects if it fails. This should help to keep
errors from vanishing.

Also changed the `bundle()` method to write the bundled output only if
successful. (Might want to reconsider this – would an empty bundled
file be better than a file left over from a previous run?)

Fixed the code that sets a default babel configuration. It was invoking
`rc` so as to include command line arguments in the output, which caused
Babel to fail.

Updated the tests to check for valid bundled output. (The old test code
reported success even though several of the tests actually failed to
produce valid results.)